refactor: reorganize Iroh transport protocol and server structure
